+++ bg_atlasapi/bg_atlas.py
@@ -1,14 +1,8 @@
 from pathlib import Path
 import tarfile
 import requests
-
 from rich import print as rprint
-from rich.table import Table
-from rich.panel import Panel
-from rich.pretty import Pretty
-from rich.text import Text
-from rich.console import Console
-from io import StringIO
+
 
 from bg_atlasapi import utils, config, core, descriptors
 
@@ -185,81 +179,3 @@ class BrainGlobeAtlas(core.Atlas):
             )
             return False
         return True
-
-    def __repr__(self):
-        """Fancy print for the atlas providing authors information."""
-        name_split = self.atlas_name.split("_")
-        pretty_name = "{} {} atlas (res. {})".format(*name_split)
-        return pretty_name
-
-    def __str__(self):
-        """
-        If the atlas metadat are to be printed
-        with the built in print function instead of rich's, then
-        print the rich panel as a string.
-
-        It will miss the colors.
-
-        """
-        buf = StringIO()
-        _console = Console(file=buf, force_jupyter=False)
-        _console.print(self)
-
-        return buf.getvalue()
-
-    def __rich_console__(self, *args):
-        """
-        Method for rich API's console protocol.
-        Prints the atlas metadata as a table nested in a panel
-        """
-        orange = "#f59e42"
-        dimorange = "#b56510"
-        gray = "#A9A9A9"
-        mocassin = "#FFE4B5"
-        cit_name, cit_link = self.metadata["citation"].split(", ")
-
-        # Create a rich table
-        tb = Table(
-            box=None,
-            show_lines=False,
-            title=self.atlas_name.replace("_", " ").capitalize(),
-            title_style=f"bold {orange}",
-        )
-
-        # Add entries to table
-        tb.add_column(
-            style=f"bold {mocassin}",
-            justify="right",
-            min_width=8,
-            max_width=40,
-        )
-        tb.add_column(min_width=20, max_width=48)
-
-        tb.add_row(
-            "name:",
-            Text.from_markup(
-                self.metadata["name"]
-                + f' [{gray}](v{self.metadata["version"]})'
-            ),
-        )
-        tb.add_row(
-            "species:", Text.from_markup(f'[i]{self.metadata["species"]}')
-        )
-        tb.add_row(
-            "citation:", Text.from_markup(f"{cit_name} [{gray}]{cit_link}")
-        )
-        tb.add_row("link:", Text.from_markup(self.metadata["atlas_link"]))
-
-        tb.add_row("")
-        tb.add_row(
-            "orientation:",
-            Text.from_markup(f"[bold]{self.metadata['orientation']}"),
-        )
-        tb.add_row("symmetric:", Pretty(self.metadata["symmetric"]))
-        tb.add_row("resolution:", Pretty(self.metadata["resolution"]))
-        tb.add_row("shape:", Pretty(self.metadata["shape"]))
-
-        # Fit into panel and yield
-        panel = Panel.fit(tb, border_style=dimorange)
-
-        yield panel
